requirements.txt
# Web & APIs
Flask>=2.0.3
fastapi>=0.95.2
uvicorn>=0.18.3
Werkzeug>=2.0.3
jinja2>=3.1.2
python-dotenv>=0.21.0
python-multipart>=0.0.6
websockets>=10.3
watchfiles>=0.17.0

# Data & ML
numpy>=1.24.0
pandas>=1.5.3
matplotlib>=3.6.3
seaborn>=0.12.2
xgboost>=1.6.2
scikit-learn>=1.2.2
imbalanced-learn>=0.10.1
nltk>=3.8.1

# Database & Cloud
boto3>=1.26.105
pymongo>=4.2.0
dnspython>=2.2.1

# Dev & Utilities
dill>=0.3.5.1
aiofiles>=23.1.0
setuptools>=65.5.0
